Update to 4.1

+++ administrator/components/com_tags/src/View/Tags/HtmlView.php
@@ -180,7 +180,7 @@ class HtmlView extends BaseHtmlView
 			}
 		}
 
-		if ($this->state->get('filter.published') == -2 && $canDo->get('core.delete'))
+		if (!$this->isEmptyState && $this->state->get('filter.published') == -2 && $canDo->get('core.delete'))
 		{
 			$toolbar->delete('tags.delete')
 				->text('JTOOLBAR_EMPTY_TRASH')
